kJoules, BikeScore, xpower displayed, fixed length speed/avg speed, refactored grid layout.
kJoules : take the avg power and multiple it by the number of seconds.

bikescore, xpower : added a rideFile object to calculate the bikescore, xpower on a 5 second
period.  It has it's own timer.

fixed display of speed and avgSpeed:
old aglorithm would not display the tenths place if it was zero.  New aglorithm
always displays the tenths place.



Signed-off-by: Greg Lonnon <greg.lonnon@gmail.com>