RealtimeWindow fix:  the disk_timer wasn't being enabled.
Looking at the history, it looks like we had a bad merge that removed
the check on the trainTool->recordSelector box to turn on the disk timer.

this checks the selector and turns it on.

Also, I removed some "dead" code at the end of an if statement.
The if condition is only followed if RT_RECORDING is set.  Thus, nulling
out the RT_RECORDING bit in the else clause is a nop.

Signed-off-by: Greg Lonnon <greg.lonnon@gmail.com>